requirements.txt
torch
torchvision
torchaudio
numpy
opencv-python
scikit-learn
matplotlib
tqdm
tensorboard
pillow
